aaaaaaaaaaaaaaaaaaaaaaa

+++ declarations.h
@@ -6,7 +6,7 @@
 /*   By: szyn <szyn@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 18:52:09 by aymel-ha          #+#    #+#             */
-/*   Updated: 2025/11/30 19:03:41 by szyn             ###   ########.fr       */
+/*   Updated: 2025/11/30 19:55:35 by szyn             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int		ft_printf(const char *str, ...);
 size_t	specifier_c(char c);
 size_t	specifier_d(int number);
 size_t	specifier_s(char *s);
-size_t	specifier_x(unsigned int n);
+size_t	specifier_x(unsigned long n);
 size_t	specifier_x_uppercase(unsigned int n);
 size_t	specifier_p(void *pointer);
 size_t	specifier_i(long n);
