+++ specifier_x.c
@@ -6,13 +6,13 @@
 /*   By: szyn <szyn@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/27 12:04:02 by aymel-ha          #+#    #+#             */
-/*   Updated: 2025/11/30 18:56:44 by szyn             ###   ########.fr       */
+/*   Updated: 2025/11/30 19:59:04 by szyn             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "declarations.h"
 
-size_t	specifier_x(unsigned int n)
+size_t	specifier_x(unsigned long n)
 {
 	size_t	len;
 	char	*p;
